SceneManager Implementation
Added Main Menu and Game Over scenes, with a singleton manager to handle the transition between them.
Allows for loading the next scene in order (would require all levels to be in order)
Allows to load a specific scene by name

TODO: Allow Levels to be added to a list which can be iterated through and doesn't require levels added to the scene in order.